book: Extend memory management chapter
