+++ book/listings/g_object_memory_management/1/main.rs
@@ -1,8 +1,6 @@
-use std::cell::Cell;
-use std::rc::Rc;
-
 use gtk::prelude::*;
-use gtk::{self, glib, Application, ApplicationWindow, Button, Orientation};
+use gtk::{glib, Application, ApplicationWindow, Button};
+use std::cell::Cell;
 
 const APP_ID: &str = "org.gtk_rs.GObjectMemoryManagement1";
 
@@ -16,7 +14,9 @@ fn main() -> glib::ExitCode {
     // Run the application
     app.run()
 }
-fn build_ui(app: &Application) {
+
+// ANCHOR: build_ui
+fn build_ui(application: &Application) {
     // Create two buttons
     let button_increase = Button::builder()
         .label("Increase")
@@ -25,38 +25,22 @@ fn build_ui(app: &Application) {
         .margin_start(12)
         .margin_end(12)
         .build();
-    let button_decrease = Button::builder()
-        .label("Decrease")
-        .margin_top(12)
-        .margin_bottom(12)
-        .margin_start(12)
-        .margin_end(12)
-        .build();
 
-    // ANCHOR: callback
-    // Reference-counted object with inner-mutability
-    let number = Rc::new(Cell::new(0));
+    // A mutable integer
+    let number = Cell::new(0);
 
-    // Connect callbacks, when a button is clicked `number` will be changed
-    let number_copy = number.clone();
-    button_increase.connect_clicked(move |_| number_copy.set(number_copy.get() + 1));
-    button_decrease.connect_clicked(move |_| number.set(number.get() - 1));
-    // ANCHOR_END: callback
-
-    // Add buttons to `gtk_box`
-    let gtk_box = gtk::Box::builder()
-        .orientation(Orientation::Vertical)
-        .build();
-    gtk_box.append(&button_increase);
-    gtk_box.append(&button_decrease);
+    // Connect callbacks
+    // When a button is clicked, `number` should be changed
+    button_increase.connect_clicked(move |_| number.set(number.get() + 1));
 
     // Create a window
     let window = ApplicationWindow::builder()
-        .application(app)
+        .application(application)
         .title("My GTK App")
-        .child(&gtk_box)
+        .child(&button_increase)
         .build();
 
     // Present the window
     window.present();
 }
+// ANCHOR_END: build_ui
