+++ book/listings/g_object_memory_management/2/main.rs
@@ -1,7 +1,6 @@
 use std::cell::Cell;
 use std::rc::Rc;
 
-use glib::clone;
 use gtk::prelude::*;
 use gtk::{self, glib, Application, ApplicationWindow, Button, Orientation};
 
@@ -34,17 +33,14 @@ fn build_ui(app: &Application) {
         .margin_end(12)
         .build();
 
-    // Reference-counted object with inner mutability
-    let number = Rc::new(Cell::new(0));
-    // Connect callbacks
-    // When a button is clicked, `number` will be changed
     // ANCHOR: callback
-    button_increase.connect_clicked(clone!(@strong number => move |_| {
-        number.set(number.get() + 1);
-    }));
-    button_decrease.connect_clicked(move |_| {
-        number.set(number.get() - 1);
-    });
+    // Reference-counted object with inner-mutability
+    let number = Rc::new(Cell::new(0));
+
+    // Connect callbacks, when a button is clicked `number` will be changed
+    let number_copy = number.clone();
+    button_increase.connect_clicked(move |_| number_copy.set(number_copy.get() + 1));
+    button_decrease.connect_clicked(move |_| number.set(number.get() - 1));
     // ANCHOR_END: callback
 
     // Add buttons to `gtk_box`
