+++ book/listings/g_object_memory_management/5/main.rs
@@ -5,7 +5,7 @@ use glib::clone;
 use gtk::prelude::*;
 use gtk::{self, glib, Application, ApplicationWindow, Button, Orientation};
 
-const APP_ID: &str = "org.gtk_rs.GObjectMemoryManagement4";
+const APP_ID: &str = "org.gtk_rs.GObjectMemoryManagement5";
 
 fn main() -> glib::ExitCode {
     // Create a new application
@@ -35,36 +35,41 @@ fn build_ui(app: &Application) {
         .margin_end(12)
         .build();
 
+    // Reference-counted object with inner mutability
     let number = Rc::new(Cell::new(0));
 
     // ANCHOR: callback
     // Connect callbacks
     // When a button is clicked, `number` and label of the other button will be changed
-    button_increase.connect_clicked(clone!(@weak number, @strong button_decrease =>
+    button_increase.connect_clicked(clone!(@weak number, @weak button_decrease =>
         move |_| {
             number.set(number.get() + 1);
             button_decrease.set_label(&number.get().to_string());
     }));
-    button_decrease.connect_clicked(clone!(@strong button_increase =>
+    button_decrease.connect_clicked(clone!(@weak button_increase =>
         move |_| {
             number.set(number.get() - 1);
             button_increase.set_label(&number.get().to_string());
     }));
     // ANCHOR_END: callback
 
+    // ANCHOR: box_append
     // Add buttons to `gtk_box`
     let gtk_box = gtk::Box::builder()
         .orientation(Orientation::Vertical)
         .build();
     gtk_box.append(&button_increase);
     gtk_box.append(&button_decrease);
+    // ANCHOR_END: box_append
 
+    // ANCHOR: window_child
     // Create a window
     let window = ApplicationWindow::builder()
         .application(app)
         .title("My GTK App")
         .child(&gtk_box)
         .build();
+    // ANCHOR_END: window_child
 
     // Present the window
     window.present();
